Refactor: make sure everything has return types
