+++ src/lib/responder.ts
@@ -3,7 +3,7 @@ const spark = require('ciscospark/env');
 import responseFactory from './response-triggers/index';
 import __logger from './logger';
 
-async function respond(messageEvent) {
+async function respond(messageEvent: {data: {id: string}}) {
 	try {
 		let message : MessageObject = await spark.messages.get(messageEvent.data.id);
 		__logger.debug('processing message: ' + JSON.stringify(message));
