+++ src/models/parsed-message.ts
@@ -1,0 +1,9 @@
+import * as xml from "libxmljs";
+
+export interface ParsedMessage {
+	fromPerson: string;
+	toPersonId: string;
+	botId: string;
+	children: xml.Element[];
+	comment: string;
+}
